Add files via upload
This is a "Talk to Speak" web application, designed to convert your text into speak. Build using  HTML for structure, CSS for an attractive and and JavaScript for seamless functionality, this app contains two button one is Listen and another one is Stop. Simply click the Listen Button for Listening the text that user entered in the Text-area. If you want to stop hearing the click the button Stop.